feat(ECDSA): conform to subtle interface
